:wrench: Change over to remove goto and recursions
As a suggestion at https://codereview.stackexchange.com/a/206238/182935
"Setup called itself if it fails to read user name or password, or username already exist This coupled with previous point blows up the stack because getline will fail which result in username or password to be empty which then recursively call setup. Because log1 etc are global variable the recursive call can modify them half way in the previous function. Also the new instance of setup will run then the previous setup finish which duplicate the work that seems to be a bug."

Bug:🐛 Login now doesn't work, signup only